+++ implementation/sipamato/src/main/java/ch/difty/sipamato/persistance/jooq/search/SearchOrderRepository.java
@@ -4,6 +4,6 @@ import ch.difty.sipamato.db.tables.records.SearchOrderRecord;
 import ch.difty.sipamato.entity.SearchOrder;
 import ch.difty.sipamato.persistance.jooq.EntityRepository;
 
-public interface SearchOrderRepository extends EntityRepository<SearchOrderRecord, SearchOrder, Integer, SearchOrderRecordMapper, SearchOrderFilter> {
+public interface SearchOrderRepository extends EntityRepository<SearchOrderRecord, SearchOrder, Long, SearchOrderRecordMapper, SearchOrderFilter> {
 
 }
